refactored choice models
