+++ src/main/python/choicemodels/estimate_mixed_plan_choice.py
@@ -2,62 +2,34 @@
 # -*- coding: utf-8 -*-
 
 import argparse
-import numpy as np
-import pandas as pd
+
 from xlogit import MixedLogit, MultinomialLogit
 from xlogit.utils import wide_to_long
 
-from estimate_plan_choice import calc_costs
+from prepare import read_plan_choices
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Estimate the plan choice mixed logit model")
-    parser.add_argument("--input", help="Path to the input file", type=str, default="../../../plan-choices.csv")
+    parser.add_argument("--input", help="Path to the input file", type=str,
+                        default="../../../plan-choices-bestK_9-tt-only.csv")
     parser.add_argument("--n-draws", help="Number of draws for the estimation", type=int, default=1500)
     parser.add_argument("--batch-size", help="Batch size for the estimation", type=int, default=None)
-    parser.add_argument("--sample", help="Use sample of choice data", type=float, default=0.2)
+    parser.add_argument("--sample", help="Use sample of choice data", type=float, default=1)
     parser.add_argument("--seed", help="Random seed", type=int, default=0)
     parser.add_argument("--mnl", help="Use MNL instead of mixed logit", action="store_true")
 
     args = parser.parse_args()
 
-    df_wide = pd.read_csv(args.input, comment="#")
+    ds = read_plan_choices(args.input, sample=args.sample, seed=args.seed)
 
-    modes = list(df_wide.columns.str.extract(r"_([a-zA-z]+)_usage", expand=False).dropna().unique())
-    print("Modes: ", modes)
-
-    k = df_wide.columns.str.extract(r"plan_(\d+)", expand=False).dropna().to_numpy(int).max()
-    print("Number of plans: ", len(df_wide))
-    print("Number of choices for plan: ", k)
-
-    # df_wide["p_id"] = df_wide["p_id"].str.replace(r"_\d+$", "", regex=True)
-    # df_wide["person"] = df_wide["person"].astype('category').cat.codes
-
-    # sample = set(df_wide.person.sample(frac=0.2))
-    # df_wide = df_wide[df_wide.person.isin(sample)]
-    if args.sample < 1:
-        df_wide = df_wide.sample(frac=args.sample, random_state=args.seed)
-
-    print("Modes:", modes)
-    print("Number of choices:", len(df_wide))
-
-    df_wide['custom_id'] = np.arange(len(df_wide))  # Add unique identifier
-    df_wide['choice'] = df_wide['choice'].map({1: "plan_1"})
-
-    df_wide = calc_costs(df_wide, k, modes)
-
-    varying = list(df_wide.columns.str.extract(r"plan_1_([a-zA-z_]+)", expand=False).dropna().unique())
-
-    print("Varying:", varying)
-
-    df = wide_to_long(df_wide, id_col='custom_id', alt_name='alt', sep='_',
-                      alt_list=[f"plan_{i}" for i in range(1, k + 1)], empty_val=0,
-                      varying=varying, alt_is_prefix=True)
+    df = wide_to_long(ds.df, id_col='custom_id', alt_name='alt', sep='_',
+                      alt_list=[f"plan_{i}" for i in range(1, ds.k + 1)], empty_val=0,
+                      varying=ds.varying, alt_is_prefix=True)
 
     MixedLogit.check_if_gpu_available()
 
-
     # ASC is present as mode_usage
-    varnames = [f"{mode}_usage" for mode in modes if mode != "walk" and mode != "car"] + ["car_used"]
+    varnames = [f"{mode}_usage" for mode in ds.modes if mode != "walk" and mode != "car"] + ["car_used"]
     # varnames += ["pt_ride_hours", "car_ride_hours", "bike_ride_hours"]
     # varnames = ["car_used", "car_usage"]
 
@@ -77,7 +49,7 @@ if __name__ == "__main__":
                   optim_method='L-BFGS-B')
 
     else:
-        #varnames += ["car_usage"]
+        # varnames += ["car_usage"]
 
         model = MultinomialLogit()
         model.fit(X=df[varnames], y=df['choice'], weights=df['weight'], varnames=varnames,
